# apps/api/api/openapi.yaml
openapi: 3.0.3
info:
  title: Do indeksa API
  version: 0.1.0
paths:
  /v1/auth/google:
    get:
      operationId: startGoogleAuth
      summary: Start browser-bound Google authorization
      parameters:
        - name: redirect
          in: query
          description: Relative path to return to after sign-in
          schema:
            type: string
            default: /
      responses:
        "302":
          description: Redirect to Google or the canonical preview bootstrap
          headers:
            Location:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/auth/google/bootstrap:
    get:
      operationId: bootstrapGoogleAuth
      summary: Bind a non-canonical sign-in to its initiating browser
      parameters:
        - name: request
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 4096
      responses:
        "302":
          description: Redirect to the next browser-binding step or Google
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/auth/google/callback:
    get:
      operationId: googleAuthCallback
      summary: OAuth callback registered with Google
      parameters:
        - name: code
          in: query
          schema:
            type: string
            maxLength: 4096
        - name: state
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 4096
        - name: error
          in: query
          description: OAuth error from Google, e.g. access_denied on cancel
          schema:
            type: string
      responses:
        "302":
          description: >-
            Canonical origin: session cookie is set and the browser is sent to
            the redirect path from state. Other allowed origin: the browser is
            sent to that origin's /api/v1/auth/exchange with a one-time code.
            Google errors (cancelled consent) return to the initiating page
            without a session.
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/auth/exchange:
    get:
      operationId: exchangeAuthCode
      summary: Trade a one-time code for a session cookie on the current origin
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: binding
          in: query
          required: true
          schema:
            type: string
            pattern: "^[A-Za-z0-9_-]{22}$"
      responses:
        "302":
          description: Session cookie set, browser sent to the stored redirect path
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session
      responses:
        "204":
          description: Session revoked, cookie cleared
          headers:
            Set-Cookie:
              schema:
                type: string
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/attempts:
    get:
      operationId: listAttempts
      summary: Task attempt journal of the current user
      responses:
        "200":
          description: Up to the most recent 1000 attempts, ordered oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attempt"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
    post:
      operationId: recordAttempts
      summary: Append attempts to the journal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 500
              items:
                $ref: "#/components/schemas/NewAttempt"
      responses:
        "204":
          description: Attempts recorded
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me:
    get:
      operationId: getMe
      summary: Current user
      responses:
        "200":
          description: Authenticated user; session cookie is refreshed when the TTL slides
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
components:
  schemas:
    Attempt:
      type: object
      required:
        - taskId
        - slot
        - correct
        - source
        - helpLevel
        - at
      properties:
        taskId:
          type: string
          maxLength: 64
          pattern: ^[a-z0-9-]+$
        slot:
          type: integer
          minimum: 1
          maximum: 10
        correct:
          type: boolean
        source:
          type: string
          enum:
            - diagnostic
            - practice
            - simulation
        helpLevel:
          type: integer
          minimum: 0
          maximum: 3
          description: "Highest help used before the attempt: 0 none, 1-2 hint level, 3 full solution"
        at:
          type: string
          format: date-time
    NewAttempt:
      type: object
      required:
        - taskId
        - slot
        - correct
        - source
      properties:
        taskId:
          type: string
          maxLength: 64
          pattern: ^[a-z0-9-]+$
        slot:
          type: integer
          minimum: 1
          maximum: 10
        correct:
          type: boolean
        source:
          type: string
          enum:
            - diagnostic
            - practice
            - simulation
        helpLevel:
          type: integer
          minimum: 0
          maximum: 3
          description: "Highest help used before the attempt: 0 none, 1-2 hint level, 3 full solution"
        at:
          type: string
          format: date-time
    User:
      type: object
      required:
        - id
        - email
        - name
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        pictureUrl:
          type: string
          format: uri
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Session mutation did not come from the externally visible request origin
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: OAuth provider is temporarily unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Internal:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"