# apps/api/internal/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package api

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for AttemptSource.
const (
	AttemptSourceDiagnostic AttemptSource = "diagnostic"
	AttemptSourcePractice   AttemptSource = "practice"
	AttemptSourceSimulation AttemptSource = "simulation"
)

// Valid indicates whether the value is a known member of the AttemptSource enum.
func (e AttemptSource) Valid() bool {
	switch e {
	case AttemptSourceDiagnostic:
		return true
	case AttemptSourcePractice:
		return true
	case AttemptSourceSimulation:
		return true
	default:
		return false
	}
}

// Defines values for NewAttemptSource.
const (
	NewAttemptSourceDiagnostic NewAttemptSource = "diagnostic"
	NewAttemptSourcePractice   NewAttemptSource = "practice"
	NewAttemptSourceSimulation NewAttemptSource = "simulation"
)

// Valid indicates whether the value is a known member of the NewAttemptSource enum.
func (e NewAttemptSource) Valid() bool {
	switch e {
	case NewAttemptSourceDiagnostic:
		return true
	case NewAttemptSourcePractice:
		return true
	case NewAttemptSourceSimulation:
		return true
	default:
		return false
	}
}

// Attempt defines model for Attempt.
type Attempt struct {
	At      time.Time `json:"at"`
	Correct bool      `json:"correct"`

	// HelpLevel Highest help used before the attempt: 0 none, 1-2 hint level, 3 full solution
	HelpLevel int           `json:"helpLevel"`
	Slot      int           `json:"slot"`
	Source    AttemptSource `json:"source"`
	TaskId    string        `json:"taskId"`
}

// AttemptSource defines model for Attempt.Source.
type AttemptSource string

// Error defines model for Error.
type Error struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// NewAttempt defines model for NewAttempt.
type NewAttempt struct {
	At      *time.Time `json:"at,omitempty"`
	Correct bool       `json:"correct"`

	// HelpLevel Highest help used before the attempt: 0 none, 1-2 hint level, 3 full solution
	HelpLevel *int             `json:"helpLevel,omitempty"`
	Slot      int              `json:"slot"`
	Source    NewAttemptSource `json:"source"`
	TaskId    string           `json:"taskId"`
}

// NewAttemptSource defines model for NewAttempt.Source.
type NewAttemptSource string

// User defines model for User.
type User struct {
	Email      openapi_types.Email `json:"email"`
	Id         openapi_types.UUID  `json:"id"`
	Name       string              `json:"name"`
	PictureUrl *string             `json:"pictureUrl,omitempty"`
}

// BadGateway defines model for BadGateway.
type BadGateway = Error

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Forbidden defines model for Forbidden.
type Forbidden = Error

// Internal defines model for Internal.
type Internal = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// RecordAttemptsJSONBody defines parameters for RecordAttempts.
type RecordAttemptsJSONBody = []NewAttempt

// ExchangeAuthCodeParams defines parameters for ExchangeAuthCode.
type ExchangeAuthCodeParams struct {
	Code    string `form:"code" json:"code"`
	Binding string `form:"binding" json:"binding"`
}

// StartGoogleAuthParams defines parameters for StartGoogleAuth.
type StartGoogleAuthParams struct {
	// Redirect Relative path to return to after sign-in
	Redirect *string `form:"redirect,omitempty" json:"redirect,omitempty"`
}

// BootstrapGoogleAuthParams defines parameters for BootstrapGoogleAuth.
type BootstrapGoogleAuthParams struct {
	Request string `form:"request" json:"request"`
}

// GoogleAuthCallbackParams defines parameters for GoogleAuthCallback.
type GoogleAuthCallbackParams struct {
	Code  *string `form:"code,omitempty" json:"code,omitempty"`
	State string  `form:"state" json:"state"`

	// Error OAuth error from Google, e.g. access_denied on cancel
	Error *string `form:"error,omitempty" json:"error,omitempty"`
}

// RecordAttemptsJSONRequestBody defines body for RecordAttempts for application/json ContentType.
type RecordAttemptsJSONRequestBody = RecordAttemptsJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Task attempt journal of the current user
	// (GET /v1/attempts)
	ListAttempts(w http.ResponseWriter, r *http.Request)
	// Append attempts to the journal
	// (POST /v1/attempts)
	RecordAttempts(w http.ResponseWriter, r *http.Request)
	// Trade a one-time code for a session cookie on the current origin
	// (GET /v1/auth/exchange)
	ExchangeAuthCode(w http.ResponseWriter, r *http.Request, params ExchangeAuthCodeParams)
	// Start browser-bound Google authorization
	// (GET /v1/auth/google)
	StartGoogleAuth(w http.ResponseWriter, r *http.Request, params StartGoogleAuthParams)
	// Bind a non-canonical sign-in to its initiating browser
	// (GET /v1/auth/google/bootstrap)
	BootstrapGoogleAuth(w http.ResponseWriter, r *http.Request, params BootstrapGoogleAuthParams)
	// OAuth callback registered with Google
	// (GET /v1/auth/google/callback)
	GoogleAuthCallback(w http.ResponseWriter, r *http.Request, params GoogleAuthCallbackParams)
	// Revoke the current session
	// (POST /v1/auth/logout)
	Logout(w http.ResponseWriter, r *http.Request)
	// Current user
	// (GET /v1/me)
	GetMe(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Task attempt journal of the current user
// (GET /v1/attempts)
func (_ Unimplemented) ListAttempts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Append attempts to the journal
// (POST /v1/attempts)
func (_ Unimplemented) RecordAttempts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Trade a one-time code for a session cookie on the current origin
// (GET /v1/auth/exchange)
func (_ Unimplemented) ExchangeAuthCode(w http.ResponseWriter, r *http.Request, params ExchangeAuthCodeParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Start browser-bound Google authorization
// (GET /v1/auth/google)
func (_ Unimplemented) StartGoogleAuth(w http.ResponseWriter, r *http.Request, params StartGoogleAuthParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Bind a non-canonical sign-in to its initiating browser
// (GET /v1/auth/google/bootstrap)
func (_ Unimplemented) BootstrapGoogleAuth(w http.ResponseWriter, r *http.Request, params BootstrapGoogleAuthParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// OAuth callback registered with Google
// (GET /v1/auth/google/callback)
func (_ Unimplemented) GoogleAuthCallback(w http.ResponseWriter, r *http.Request, params GoogleAuthCallbackParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Revoke the current session
// (POST /v1/auth/logout)
func (_ Unimplemented) Logout(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Current user
// (GET /v1/me)
func (_ Unimplemented) GetMe(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListAttempts operation middleware
func (siw *ServerInterfaceWrapper) ListAttempts(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListAttempts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RecordAttempts operation middleware
func (siw *ServerInterfaceWrapper) RecordAttempts(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RecordAttempts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ExchangeAuthCode operation middleware
func (siw *ServerInterfaceWrapper) ExchangeAuthCode(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params ExchangeAuthCodeParams

	// ------------- Required query parameter "code" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "code", r.URL.Query(), &params.Code, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "code"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "code", Err: err})
		}
		return
	}

	// ------------- Required query parameter "binding" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "binding", r.URL.Query(), &params.Binding, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "binding"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "binding", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ExchangeAuthCode(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// StartGoogleAuth operation middleware
func (siw *ServerInterfaceWrapper) StartGoogleAuth(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params StartGoogleAuthParams

	// ------------- Optional query parameter "redirect" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "redirect", r.URL.Query(), &params.Redirect, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "redirect"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "redirect", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.StartGoogleAuth(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// BootstrapGoogleAuth operation middleware
func (siw *ServerInterfaceWrapper) BootstrapGoogleAuth(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params BootstrapGoogleAuthParams

	// ------------- Required query parameter "request" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "request", r.URL.Query(), &params.Request, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "request"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "request", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.BootstrapGoogleAuth(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GoogleAuthCallback operation middleware
func (siw *ServerInterfaceWrapper) GoogleAuthCallback(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GoogleAuthCallbackParams

	// ------------- Optional query parameter "code" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "code", r.URL.Query(), &params.Code, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "code"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "code", Err: err})
		}
		return
	}

	// ------------- Required query parameter "state" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "state", r.URL.Query(), &params.State, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "state"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "state", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "error" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "error", r.URL.Query(), &params.Error, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "error"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "error", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GoogleAuthCallback(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Logout(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetMe operation middleware
func (siw *ServerInterfaceWrapper) GetMe(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMe(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/attempts", wrapper.ListAttempts)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/attempts", wrapper.RecordAttempts)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/auth/exchange", wrapper.ExchangeAuthCode)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/auth/google", wrapper.StartGoogleAuth)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/auth/google/bootstrap", wrapper.BootstrapGoogleAuth)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/auth/google/callback", wrapper.GoogleAuthCallback)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/auth/logout", wrapper.Logout)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/me", wrapper.GetMe)
	})

	return r
}